+++ README.md
@@ -1,11 +1,11 @@
 # YandexDiplomaProject
 # ссылка на мой репозиторий https://github.com/DmJavaScript/YandexDiplomaProject
 
-1. Сайт "NewsAnalizer" Актуальная версия программы v.0.1.7 от 29 декабря 2019г.
+1. Сайт "NewsAnalizer" Актуальная версия программы v.0.2.2 от 19 января 2020г.
 
 Данная версия содержит код проектной дипломной работы в рамках курса по изучению языков программирования HTML, CSS, Javascript
 Для проверки какими технологиями овладел студент, в качестве макета, сверстан 3-х страничный сайт. Новостной поисковик с обращением к внешнему api, с адаптивной версткой.
-Последнее обновление: добавлен JavaScript для всех страниц.
+Последнее обновление: добавлен модульный JavaScript для всех страниц.
 
 
 # Инструкции по развёртыванию:
@@ -38,7 +38,8 @@
     "@glidejs/glide": "3.4.1",
     "babel-polyfill": "6.26.0",
     "image-webpack-loader": "6.0.0",
-    "webpack-dev-server": "3.9.0"
+    "webpack-dev-server": "3.10.1"
+
 
 
 # можете встроить этот код в свой проект, для этого достаточно загрузить файлы архивом или:
