+++ src/about.html
@@ -61,7 +61,7 @@
 
       <section class='commit-history'>
         <h2 class='commit-history__title'>История коммитов из&nbsp;GitHub</h2>
-        <a href='https://github.com/DmJavaScript/YandexDiplomaProject/commits/develop' class='commit-history__external-link' target='_blank'>Открыть в&nbsp;Github&nbsp;&nbsp;&#x276F;</a>
+        <a href='https://github.com/DmJavaScript/YandexDiplomaProject/commits' class='commit-history__external-link' target='_blank'>Открыть в&nbsp;Github&nbsp;&nbsp;&#x276F;</a>
         <div id='options-focus-at' class='slider glide commit-history__slider'>
           <div class='slider__track glide__track' data-glide-el='track'>
             <div class='slider__slides glide__slides'>
