+++ src/blocks/slider/__slides/slides_list.js
@@ -1,5 +1,5 @@
-import {createSlide} from './constants.js';
-import {formatDate} from './utils.js';
+import {createSlide, glideConfig, glide} from '../../../pages/about.js';
+import {formatDate} from '../../../js/utils.js';
 
 export class SlidesList {
   constructor() {
@@ -17,9 +17,10 @@ export class SlidesList {
 
   render(data) { //метод для автоматической отрисовки карточек из списка addCard
     data.forEach((e, i) => {
-      this._currentNum = i; // this._currentNum =  `=` + `${i}`;
+      this._currentNum =  `=` + `${i}`;
       this._addCard(formatDate(data[i].commit.author.date), data[i].author.avatar_url, data[i].commit.author.name, data[i].commit.author.email, data[i].commit.message, this._currentNum);
     });
+    glide.mount();
   }
 
 }
