+++ src/js/storage_data.js
@@ -1,4 +1,4 @@
-import {regExpHTTPLinkFirst, cardsList, input} from '../js/constants.js';
+import {regExpHTTPLinkFirst, cardsList, input} from '../pages/index.js';
 import {formatDate} from '../js/utils.js';
 
 export class StorageData {
@@ -8,8 +8,8 @@ export class StorageData {
   }
 
   chekingStorage () {// Метод получения массива новостей с локального хранилища
-    const _recivedData = JSON.parse(localStorage.getItem("NewsApiLocalStorage"));
-    const _recivedDataRequest = localStorage.getItem("NewsApiRequest");
+    const _recivedData = JSON.parse(localStorage.getItem('NewsApiLocalStorage'));
+    const _recivedDataRequest = localStorage.getItem('NewsApiRequest');
     if (_recivedDataRequest !== null && _recivedData !== null) {
       this._input.value = _recivedDataRequest;
       let resultsArray = [];
@@ -21,7 +21,7 @@ export class StorageData {
   }
 
   _preparePackageArray (resultsArray) {
-    const _descriptionTextPreview = resultsArray.description; //.replace(regExpHTTPLinkFirst, ''); //временно отложил решение
+    const _descriptionTextPreview = resultsArray.description; //.replace(regExpHTTPLinkFirst, ''); //временно отложил решение(необязательная для диплома функция)
     const _cardDate = formatDate(resultsArray.publishedAt);
     const _dateTime = resultsArray.publishedAt;
     return [_dateTime, resultsArray.url, resultsArray.urlToImage, _cardDate, resultsArray.title, _descriptionTextPreview, resultsArray.source.name];
