+++ src/pages/about.js
@@ -1,8 +1,22 @@
 import "./about.css";
-import {githubApi, slidesList, glide} from '../js/constants.js';
+import Glide from '@glidejs/glide';
+import {glideConfig} from '../js/glide_config.js';
+
+import {GithubApi} from '../js/github_api.js';
+import {Slide} from '../blocks/slider/__card/slide.js';
+import {SlidesList} from '../blocks/slider/__slides/slides_list.js';
+
+export const githubApi = new GithubApi ({
+  baseUrl: 'https://api.github.com/repos/DmJavaScript/YandexDiplomaProject/commits',
+  headers: {
+    authorization: 'ca68fe15b70347d75428771071b8dfe19e1a8ba5',
+    'Content-Type': 'application/json'
+  }
+});
+export const createSlide = (...args) => new Slide(...args);
+export const slidesList = new SlidesList();
+export const glide = new Glide(glideConfig.container, glideConfig.settings);
 
 githubApi.getSlidesData()
           .then(data => slidesList.render(data))
           .catch((err) => console.log('Ошибка. Запрос не выполнен: ', err));
-
-glide.mount();
