+++ src/pages/details.js
@@ -1,5 +1,9 @@
 import "./details.css";
-import {shortStatistics, table} from "../js/constants.js";
+export {currentDate, sevenDaysAgo, recivedData, recivedDataRequest, regExpRequest} from '../js/constants.js';
+import {ShortStatistics} from '../blocks/header/short_statistics.js';
+import {Table} from '../blocks/search-analitics/table.js';
+export const shortStatistics = new ShortStatistics;
+export const table = new Table;
 
 shortStatistics.render();
 table.updateData();
